Tests/LibWeb/Text/expected/HTML/media-source-remove-source-buffer-and-script-blocking.txt

script.blocking instanceof DOMTokenList: true
script blocking attribute: render
sourceBuffers before remove: 1
sourceBuffers after remove: 0
second remove threw: NotFoundError
